finalize checkout

+++ internal/model/item/model.go
@@ -1,8 +1,8 @@
 package item
 
 type Item struct {
-	ID       string  `db:"id"`
-	Name     string  `db:"name"`
-	Quantity int32   `db:"quantity"`
-	Price    float64 `db:"price"`
+	ID       string `db:"id"`
+	Name     string `db:"name"`
+	Quantity int32  `db:"quantity"`
+	Price    int32  `db:"price"`
 }
